+++ R/utilities.R
@@ -724,14 +724,6 @@ degrees = function(rad) {
 
 
 
-# deprecation helpers ####
-
-# determine if a function has been called directly
-direct_call = function() {
-  ifelse(sys.nframe() == 2L, yes = TRUE, no = FALSE)
-}
-
-